Changed deserializer and misc stuff

+++ Utils/ENodeGroup.h
@@ -6,7 +6,7 @@
 struct ENodeGroup {
     uint32_t startNodeIndex;
     uint8_t nodeCount;
-    uint8_t boardPinIndex;
+    uint16_t boardPinIndex;
     bool enabled;
     IOTypes::Types type;
